add html parsing capability
