+++ Pipfile
@@ -4,8 +4,7 @@ verify_ssl = true
 name = "pypi"
 
 [packages]
-archieml = "==0.3.4"
-astroid = "==1.5.3"
+archieml = {git = "https://github.com/dailybruin/archieml-python"}
 "boto3" = "==1.4.5"
 botocore = "==1.5.92"
 certifi = "==2017.7.27.1"
@@ -13,7 +12,7 @@ chardet = "==3.0.4"
 cloudflare = "==2.1.0"
 configparser = "==3.5.0"
 defusedxml = "==0.5.0"
-django-allauth = "==0.32.0"
+django-allauth = "*"
 django-cors-headers = "==2.2.0"
 django-environ = "==0.4.4"
 django-extensions = "==1.8.1"
@@ -26,7 +25,6 @@ elasticsearch-dsl = "==6.1.0"
 "enum34" = "==1.1.6"
 future = "==0.16.0"
 futures = "==3.1.1"
-google-api-python-client = "==1.6.5"
 gunicorn = "==19.7.1"
 "httplib2" = "==0.10.3"
 idna = "==2.5"
@@ -42,7 +40,6 @@ olefile = "==0.44"
 "psycopg2" = "==2.7.3"
 "pyasn1" = "==0.4.2"
 "pyasn1-modules" = "==0.2.1"
-pylint = "==1.7.2"
 python-dateutil = "==2.6.1"
 python-openid = "==2.2.5"
 "python3-openid" = "==3.1.0"
@@ -58,14 +55,20 @@ uritemplate = "==3.0.0"
 whitenoise = "==3.3.1"
 wrapt = "==1.10.10"
 "backports.functools_lru_cache" = "==1.4"
-Django = "==2.0.2"
+Django = "==2.1"
 Pillow = "==4.2.1"
 PyDrive = "==1.3.1"
 PyYAML = "==3.12"
-monkeytype = "*"
+arrow = "*"
+bleach = "*"
+"html5lib" = "*"
 
 [dev-packages]
 mypy = "*"
+pylint-celery = "*"
+pylint-django = "*"
+monkeytype = "*"
+pylint = "*"
 
 [requires]
 python_version = "3.6"
