+++ kerckhoff/packages/constants.py
@@ -1,0 +1,7 @@
+GOOGLE_API_PREFIX = "https://www.googleapis.com/drive"
+PLAIN_TEXT = 'PT'
+HTML = 'HL'
+CONTENT_TYPE_CHOICES = (
+    (PLAIN_TEXT, 'Plaintext'),
+    (HTML, 'HTML')
+)
