+++ kerckhoff/packages/models.py
@@ -1,38 +1,28 @@
-from django.db import models
-from django.contrib.postgres.fields import JSONField, ArrayField
-from django.conf import settings
-from django.core.exceptions import ValidationError
-from django.contrib.auth.models import User
-from allauth.socialaccount.models import SocialToken
-from requests_oauthlib import OAuth2Session
-from django.utils import timezone
-from datetime import timedelta
-from PIL import Image
-import tempfile
-import boto3
-import hashlib
-from botocore.client import Config
-import boto3
-import pathlib
-import imghdr
 import re
-import requests
-import archieml
-import arrow
-from search.indexes import PackageIndex
 import logging
 
-logger = logging.getLogger(settings.APP_NAME)
+import requests
+from django.conf import settings
+from django.contrib.auth.models import User
+from django.contrib.postgres.fields import JSONField
+from django.core.exceptions import ValidationError
+from django.db import models
+from django.utils import timezone
+from requests_oauthlib import OAuth2Session
 
-S3_BUCKET = settings.S3_ASSETS_UPLOAD_BUCKET
-s3 = boto3.client('s3', 'us-west-2', config=Config(s3={'addressing_style': 'path'}))
-PREFIX = "https://www.googleapis.com/drive"
-IMAGE_REGEX = re.compile(r"!\[[^\]]+\]\(([^)]+)\)")
+from search.indexes import PackageIndex
+
+from .utils import transfer_to_s3, rewrite_image_url
+from .google_drive_actions import get_file, get_oauth2_session, list_folder, create_package, add_to_repo_folder
+from .constants import *
+
+logger = logging.getLogger(settings.APP_NAME)
 
 class PackageSet(models.Model):
     slug = models.SlugField(max_length=32, primary_key=True)
     drive_folder_id = models.CharField(max_length=512, blank=True)
     drive_folder_url = models.URLField()
+    default_content_type = models.CharField(max_length=2, choices=CONTENT_TYPE_CHOICES, default=PLAIN_TEXT)
 
     def as_dict(self):
         return {
@@ -71,6 +61,7 @@ class PackageSet(models.Model):
                 print("%s failed with error: %s" % (instance.slug, e))
                 continue
     
+
 class Package(models.Model):
     slug = models.CharField(max_length=64, primary_key=True)
     description = models.TextField(blank=True)
@@ -84,10 +75,17 @@ class Package(models.Model):
     publish_date = models.DateField()
     last_fetched_date = models.DateTimeField(null=True, blank=True)
     package_set = models.ForeignKey(PackageSet, on_delete=models.PROTECT)
-    
-    # Versioning
-    latest_version = models.ForeignKey('PackageVersion', related_name='versions', on_delete=models.CASCADE, null=True)
+    _content_type = models.CharField(max_length=2, choices=CONTENT_TYPE_CHOICES, blank=True, default="")
 
+    # Versioning
+    latest_version = models.ForeignKey('PackageVersion', related_name='versions', on_delete=models.CASCADE, null=True, blank=True)
+
+    @property
+    def content_type(self):
+        if self._content_type == "":
+            return self.package_set.default_content_type
+        else:
+            return self._content_type
 
     # For versioning feature, accepts string arguments name(of creater) and change_summary
     def create_version(self, user, change_summary):
@@ -193,199 +191,3 @@ class PackageVersion(models.Model):
     version_description = models.TextField(blank=True)
     article_data = models.TextField(blank=True)
     data = JSONField(blank=True, default=dict)
-    
-
-def rewrite_image_url(package):
-    def replace_url(fn):
-        #print(fn.group(1).replace)
-        if fn.group(1) in package.images["s3"]:
-            return fn.group().replace(fn.group(1), package.images["s3"][fn.group(1)]["url"])
-        else:
-            return fn.group()
-    text = IMAGE_REGEX.sub(replace_url,package.cached_article_preview)
-    return text
-
-def transfer_to_s3(session, package):
-    if package.images.get("s3") is None:
-        package.images["s3"] = {}
-
-    for idx, image in enumerate(package.images["gdrive"]):
-        # When the image was last modified on google drive
-        last_modified_date = arrow.get(image['modifiedDate']).datetime
-        if package.last_fetched_date is not None and package.last_fetched_date > last_modified_date:
-            logger.info(f"{ image['title'] } has not been modified since last fetch.")
-            continue
-        else:
-            logger.info(f"{ image['title'] } has been modified on Google Drive, updating.")
-
-        req = get_file(session, image["id"], download=True)
-        max_size = (1024, 1024)
-
-        with tempfile.SpooledTemporaryFile(mode="w+b") as tf:
-            print(image["title"] + " | Processing image %d ..." % idx)
-            req.raw.decode_content = True
-            header = req.raw.read(100)
-            ext = imghdr.what(None, h=header)
-            print(image["title"] + " | Found: {0}".format(ext))
-            if ext == None:
-                print("No image file found ... Continuing")
-                continue
-
-            tf.write(header + req.raw.read())
-            im = Image.open(tf)
-            image_hash = hashlib.md5(im.tobytes()).hexdigest()
-            if image["title"] in package.images["s3"]:
-                print("Old Hash found: " + package.images["s3"][image["title"]].get("hash"))
-                print("Current Hash: " + image_hash)
-                if package.images["s3"][image["title"]].get("hash") == image_hash:
-                    print(image["title"] + " in package " + package.slug + " has not been edited. Ignoring.")
-                    continue
-
-            with tempfile.SpooledTemporaryFile(mode="w+b") as wtf:
-                im.thumbnail(max_size, Image.ANTIALIAS)
-                im.save(wtf, format=im.format, optimize=True, quality=85)
-                original_fn = pathlib.Path(image["title"])
-                fn = "images/{0}/{1}-{2}{3}".format(package.slug, original_fn.stem, image_hash, original_fn.suffix)
-                wtf.seek(0)
-                response = s3.put_object(
-                    Bucket=S3_BUCKET,
-                    Key=fn,
-                    Body=wtf,
-                    ACL='public-read'
-                )
-
-                package.images["s3"][image["title"]] = {
-                    "url": "https://assets.dailybruin.com/{0}".format(fn), # TODO: replace with something configurable
-                    "key": fn,
-                    "hash": image_hash,
-                    "s3_fields": response
-                }
-    return package
-
-
-def add_to_repo_folder(session, package):
-    payload = {
-        "id": settings.REPOSITORY_FOLDER_ID
-    }
-    res = session.post(PREFIX + "/v2/files/" + package.drive_folder_id + "/parents", json=payload)
-    return res.json()
-
-def get_file(session, file_id, download=False):
-    res = session.get(PREFIX + "/v2/files/" + file_id + ("?alt=media" if download else ""), stream=download)
-    res.raise_for_status()
-    if download:
-        # Returns the requests object
-        return res
-    else:
-        return res.json()
-
-def list_folder(session, package):
-    text = "### No article document was found in this package!\n"
-    images = []
-
-    payload = {
-        "q": "'%s' in parents" % package.drive_folder_id,
-        "maxResults": 1000
-    }
-    # we assume there's always less than 100 files in a package. change this if assumption untrue
-
-    def img_check(item: dict):
-        valid_extensions = [".jpeg", ".png", ".jpg", ".gif", ".webp"]
-        for ext in valid_extensions:
-            if ext in item["title"].lower():
-                return True
-        return False
-
-    res = session.get(PREFIX + "/v2/files", params=payload)
-    items = res.json()['items']
-    article = list(filter(lambda f: "article" in f['title'], items))
-    data_files = list(filter(lambda f: ".aml" in f['title'], items))
-    images = list(filter(img_check, items))
-    folders = list(filter(lambda f: f["mimeType"] == "application/vnd.google-apps.folder", items))
-    #print("RES:")
-    #print(article)
-    #print(images)
-
-    aml_data = {}
-
-    # adds title of article as key, and parsed data as value. Saves info to aml_data
-
-    for aml in data_files:
-        if aml['mimeType'] != "application/vnd.google-apps.document":
-            req = get_file(session, aml['id'], download=True)
-            text = req.content.decode('utf-8')
-        else:
-            data = session.get(PREFIX + "/v2/files/" + aml['id'] + "/export", params={"mimeType": "text/plain"})
-            text = data.content.decode('utf-8')
-        #print("IN ARCHIEML ")
-        aml_data[aml['title']] = archieml.loads(text)
-
-    # only taking the first one - assuming there's only one article file
-    if len(article) >= 1:
-        if article[0]['mimeType'] != "application/vnd.google-apps.document":
-            req = get_file(session, article[0]['id'], download=True)
-            text = req.content.decode('utf-8')
-        else:
-            data = session.get(PREFIX + "/v2/files/" + article[0]['id'] + "/export", params={"mimeType": "text/plain"})
-            text = data.content.decode('utf-8')
-        # fix indentation for yaml
-        text = text.replace("\t", "  ")
-    # this will take REALLY long.
-
-    # return everything
-    return text, images, folders, aml_data
-
-def create_package(session, package, existing=False):
-    payload = {
-        'parents': [{'id': settings.REPOSITORY_FOLDER_ID}],
-        'title': package.slug,
-        'description': package.description,
-        'mimeType': "application/vnd.google-apps.folder"
-    }
-
-    res = session.post(PREFIX + "/v2/files", json=payload)
-    folder_resource = res.json()
-
-    file_payload = {
-        'parents': [{'id': folder_resource['id']}],
-        'title': "article.md",
-        'description': "Article data",
-        'mimeType': "application/vnd.google-apps.document"
-    }
-    session.post(PREFIX + "/v2/files", json=file_payload)
-
-    return (folder_resource['alternateLink'], folder_resource['id'])
-
-
-# https://github.com/pennersr/django-allauth/issues/420
-def get_oauth2_session(user) -> OAuth2Session:
-    """ Create OAuth2 session which autoupdates the access token if it has expired """
-
-    refresh_token_url = "https://accounts.google.com/o/oauth2/token"
-
-    social_token = SocialToken.objects.get(account__user=user, account__provider='google')
-
-    def token_updater(token):
-        social_token.token = token['access_token']
-        social_token.token_secret = token['refresh_token']
-        social_token.expires_at = timezone.now() + timedelta(seconds=int(token['expires_in']))
-        social_token.save()
-
-    client_id = social_token.app.client_id
-    client_secret = social_token.app.secret
-
-    extra = {
-        'client_id': client_id,
-        'client_secret': client_secret
-    }
-
-    expires_in = (social_token.expires_at - timezone.now()).total_seconds()
-    token = {
-        'access_token': social_token.token,
-        'refresh_token': social_token.token_secret,
-        'token_type': 'Bearer',
-        'expires_in': expires_in  # Important otherwise the token update doesn't get triggered.
-    }
-
-    return OAuth2Session(client_id, token=token, auto_refresh_kwargs=extra,
-                         auto_refresh_url=refresh_token_url, token_updater=token_updater)
