+++ requirements.txt
@@ -1,17 +1,19 @@
-archieml==0.3.4
+git+https://github.com/dailybruin/archieml-python#egg=archieml
 arrow==0.12.1
-astroid==1.6.4
-backports.functools-lru-cache==1.5
-boto3==1.7.24
-botocore==1.10.24
-certifi==2018.4.16
+astroid==2.0.4
+backports.functools-lru-cache==1.4
+bleach==3.0.2
+boto3==1.4.5
+botocore==1.5.92
+cachetools==2.1.0
+certifi==2017.7.27.1
 chardet==3.0.4
-click==6.7
+Click==7.0
 cloudflare==2.1.0
 configparser==3.5.0
 defusedxml==0.5.0
-Django==2.0.2
-django-allauth==0.32.0
+Django==2.1
+django-allauth==0.38.0
 django-cors-headers==2.2.0
 django-environ==0.4.4
 django-extensions==1.8.1
@@ -24,8 +26,11 @@ elasticsearch-dsl==6.1.0
 enum34==1.1.6
 future==0.16.0
 futures==3.1.1
-google-api-python-client==1.6.5
+google-api-python-client==1.7.4
+google-auth==1.5.1
+google-auth-httplib2==0.0.3
 gunicorn==19.7.1
+html5lib==1.0.1
 httplib2==0.10.3
 idna==2.5
 ipaddress==1.0.22
@@ -34,17 +39,21 @@ jmespath==0.9.3
 jsonlines==1.2.0
 lazy-object-proxy==1.3.1
 mccabe==0.6.1
-MonkeyType==18.5.1
-mypy==0.600
+MonkeyType==18.8.0
+mypy==0.630
+mypy-extensions==0.4.1
 oauth2client==4.1.2
 oauthlib==2.0.2
 olefile==0.44
-Pillow==5.3.0
+Pillow==4.2.1
 psycopg2==2.7.3
 pyasn1==0.4.2
 pyasn1-modules==0.2.1
 PyDrive==1.3.1
-pylint==1.7.2
+pylint==2.1.1
+pylint-celery==0.3
+pylint-django==2.0.2
+pylint-plugin-utils==0.4
 python-dateutil==2.6.1
 python-openid==2.2.5
 python3-openid==3.1.0
@@ -60,5 +69,6 @@ six==1.10.0
 typed-ast==1.1.0
 uritemplate==3.0.0
 urllib3==1.22
+webencodings==0.5.1
 whitenoise==3.3.1
 wrapt==1.10.10
